refactoring constants package name
