+++ constants/constants.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-package settings
+package constants
 
 import "time"
