+++ internal/powerplatform/api/api_client.go
@@ -17,7 +17,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-log/tflog"
-	settings "github.com/microsoft/terraform-provider-power-platform/constants"
+	"github.com/microsoft/terraform-provider-power-platform/constants"
 	config "github.com/microsoft/terraform-provider-power-platform/internal/powerplatform/config"
 	helpers "github.com/microsoft/terraform-provider-power-platform/internal/powerplatform/helpers"
 )
@@ -123,7 +123,7 @@ func (client *ApiClient) Execute(ctx context.Context, method, url string, header
 	}
 
 	//TODO: once we add timeout logic to all resource/datasource operations, we use that timeout here
-	ctx, cancel := context.WithTimeout(ctx, settings.DEFAULT_REQUEST_TIMEOUT_IN_MINUTES)
+	ctx, cancel := context.WithTimeout(ctx, constants.DEFAULT_REQUEST_TIMEOUT_IN_MINUTES)
 	defer cancel()
 
 	var response *ApiHttpResponse = nil
